Move list (business logic) to the engine package
The engine lists is a part of the domain so it should have been in the
engine package since the beginning.
